// app/store/types/projects.ts
export enum Actions {
  GET_CIW = 'GET_CIW',
  GET_CIW_REQUEST = 'GET_CIW_REQUEST',
  GET_CIW_SUCCESS = 'GET_CIW_SUCCESS',
  GET_CIW_FAILURE = 'GET_CIW_FAILURE',
  CREATE_CIW = 'CREATE_CIW',
  UPDATE_CIW = 'UPDATE_CIW',
  DELETE_CIW = 'DELETE_CIW',
  CREATE_TEMPLATE = 'CREATE_TEMPLATE',
  DELETE_TEMPLATE = 'DELETE_TEMPLATE',
}